# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel AI OS Core API
  version: 0.2.3-pilot.6
  description: >-
    PRODUCT-V1.2 / TECH-V0.2 Sprint 2.1 UAT-candidate contract. Production requests use a
    trusted JWT/SSO bearer token; roles, permissions, organization scope and
    active position assignments are resolved by the server. X-Tenant-Id and
    X-Actor-Id are accepted only when DEV_HEADER_AUTH_ENABLED=true for local
    integration. Client-supplied role or organization-scope headers are never
    trusted.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
  - TenantHeader: []
    ActorHeader: []
...
paths:
...
  /api/v1/work-duty-periods:
    get:
      tags: [Work Package]
      operationId: listWorkDutyPeriods
      summary: List duty periods
      parameters:
        - { name: positionAssignmentId, in: query, schema: { type: string, format: uuid } }
        - { name: targetOrgUnitId, in: query, schema: { type: string, format: uuid } }
        - { name: businessDate, in: query, schema: { type: string, format: date } }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Work Package]
      operationId: createWorkDutyPeriod
      summary: Create an assignment duty period
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkDutyPeriod' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
...
  /api/v1/work-data/forms:
    get:
      tags: [Work Data]
      operationId: listWorkForms
      summary: List position forms and latest versions
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Work Data]
      operationId: createWorkForm
      summary: Create a position form definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateForm' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-data/forms/{formId}/versions:
    post:
      tags: [Work Data]
      operationId: createWorkFormVersion
      summary: Create a JSON Schema form version
      parameters:
        - $ref: '#/components/parameters/FormId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateFormVersion' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-data/forms/{formId}/versions/{versionId}/publish:
    post:
      tags: [Work Data]
      operationId: publishWorkFormVersion
      summary: Publish an immutable form version
      parameters:
        - $ref: '#/components/parameters/FormId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-data/records:
    get:
      tags: [Work Data]
      operationId: listWorkRecords
      summary: List work records in effective data scope
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Work Data]
      operationId: createWorkRecord
      summary: Create a draft or submitted work record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitWorkRecord' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/work-data/records/{recordId}:
    get:
      tags: [Work Data]
      operationId: getWorkRecord
      summary: Get one work record with review and work-package linkage
      parameters:
        - $ref: '#/components/parameters/RecordId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
    put:
      tags: [Work Data]
      operationId: updateWorkRecordDraft
      summary: Update a draft using expectedVersion
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateWorkRecordDraft' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/work-data/records/{recordId}/actions/submit:
    post:
      tags: [Work Data]
      operationId: submitWorkRecordDraft
      summary: Submit a draft using expectedVersion
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitWorkRecordDraft' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/work-data/records/{recordId}/actions/review:
    post:
      tags: [Work Data]
      operationId: reviewWorkRecord
      summary: Approve or reject a submitted work record
      description: >-
        Requires work-record.review and organization scope. The submitting account cannot
        review its own record. reason is mandatory when outcome is REJECTED.
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewWorkRecord' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/work-data/records/{recordId}/supplements:
    post:
      tags: [Work Data]
      operationId: addWorkRecordSupplement
      summary: Append an immutable statement to a submitted work record
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddWorkRecordSupplement' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/work-data/records/{recordId}/attachments:
    get:
      tags: [Work Data]
      operationId: listWorkRecordAttachments
      summary: List stored image attachments for a visible work record
      description: Requires work-record.read and organization-scope access to the work record.
      parameters:
        - $ref: '#/components/parameters/RecordId'
      responses:
        '200':
          description: Attachment metadata in creation order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StoredAttachment' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Work Data]
      operationId: addWorkRecordAttachment
      summary: Register attachment metadata without uploading binary content
      description: Legacy metadata-only endpoint. Use the multipart upload endpoint for UAT image evidence.
      deprecated: true
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddAttachment' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/work-data/records/{recordId}/attachments/upload:
    post:
      tags: [Work Data]
      operationId: uploadWorkRecordImage
      summary: Upload and persist a JPEG or PNG work-record attachment
      description: >-
        Requires work-record.submit and access to the exact submitting assignment (or
        work-record.submit-for-other). The service validates actual image content,
        accepts image/jpeg and image/png, enforces the configured size limit (10 MiB by
        default), stores the binary and computes SHA-256. In formal mode malware scanning
        is fail-closed: a missing or unavailable scanner, timeout, non-zero exit, or rejected
        file fails the upload without creating attachment metadata. CLEAN means the configured
        scanner actually returned success. BYPASSED_DEV may only be produced and downloaded
        when development-header authentication is explicitly enabled. TECH-V0.2 RC exercised
        this contract with ClamAV 1.5.3 and a signed CVD database. Approved or rejected records
        are immutable.
      parameters:
        - $ref: '#/components/parameters/RecordId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: image/jpeg, image/png
      responses:
        '201':
          description: Stored attachment metadata and content URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadedAttachment' }
        '400':
          description: Invalid image or malware scanner unavailable, timed out, failed, or rejected the file
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/work-data/attachments/{attachmentId}/content:
    get:
      tags: [Work Data]
      operationId: downloadWorkRecordAttachment
      summary: Read stored work-record image content inline
      description: Requires work-record.read and organization-scope access to the owning work record.
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '200':
          description: Original JPEG or PNG content with an inline Content-Disposition filename
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
            image/png:
              schema: { type: string, format: binary }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/work-data/records/{recordId}/attachments/{attachmentId}:
    delete:
      tags: [Work Data]
      operationId: deleteWorkRecordAttachment
      summary: Delete a stored image attachment from a mutable work record
      description: >-
        Requires work-record.submit and access to the submitting assignment (or
        work-record.submit-for-other). Approved or rejected records are immutable.
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '204':
          description: Attachment metadata and binary content deleted
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/templates:
    get:
      tags: [Enterprise Template]
      operationId: listEnterpriseTemplates
      summary: List task and hotel-dashboard templates visible to the caller
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [TASK, HOTEL_DASHBOARD] }
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
    post:
      tags: [Enterprise Template]
      operationId: createEnterpriseTemplate
      summary: Create a CEO-managed template and its first draft version
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEnterpriseTemplate' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/templates/{templateId}:
    get:
      tags: [Enterprise Template]
      operationId: getEnterpriseTemplate
      summary: Get template metadata and all immutable versions
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/templates/{templateId}/versions:
    post:
      tags: [Enterprise Template]
      operationId: createEnterpriseTemplateVersion
      summary: Create a new editable draft from a published template
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEnterpriseTemplateVersion' }
      responses:
        '201': { $ref: '#/components/responses/ObjectResponse' }
  /api/v1/templates/{templateId}/versions/{versionId}:
    put:
      tags: [Enterprise Template]
      operationId: updateEnterpriseTemplateVersion
      summary: Update an editable template draft using optimistic locking
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEnterpriseTemplateVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/templates/{templateId}/versions/{versionId}/publish:
    post:
      tags: [Enterprise Template]
      operationId: publishEnterpriseTemplateVersion
      summary: Publish a template version and retire the preceding published version
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishEnterpriseTemplateVersion' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Trusted production JWT or SSO access token.
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
    ActorHeader:
      type: apiKey
      in: header
      name: X-Actor-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
  parameters:
...
    TemplateId: { name: templateId, in: path, required: true, schema: { type: string, format: uuid } }
...
    VersionId: { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
    FormId: { name: formId, in: path, required: true, schema: { type: string, format: uuid } }
    RecordId: { name: recordId, in: path, required: true, schema: { type: string, format: uuid } }
    AttachmentId: { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    ListResponse:
      description: List response
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
    ObjectResponse:
      description: Object response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: Permission denied or organization scope violation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Optimistic version or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    CreateWorkDutyPeriod:
      type: object
      required: [positionAssignmentId, targetOrgUnitId, businessDate, periodType, plannedStartAt, plannedEndAt]
      properties:
        positionAssignmentId: { type: string, format: uuid }
        targetOrgUnitId: { type: string, format: uuid }
        businessDate: { type: string, format: date }
        periodType: { type: string }
        shiftCode: { type: [string, 'null'] }
        plannedStartAt: { type: string, format: date-time }
        plannedEndAt: { type: string, format: date-time }
        sourceRecordId: { type: [string, 'null'] }
...
    CreateForm:
      type: object
      required: [code, name, formType]
      properties:
        code: { type: string }
        name: { type: string }
        formType: { type: string, enum: [DAILY_WORK, INSPECTION, COMPLAINT, EXCEPTION, FEEDBACK] }
        positionId: { type: [string, 'null'], format: uuid }
    CreateFormVersion:
      type: object
      required: [jsonSchema]
      properties:
        jsonSchema: { type: object, additionalProperties: true }
        uiSchema: { type: [object, 'null'], additionalProperties: true }
    SubmitWorkRecord:
      type: object
      required: [orgUnitId, employeeId, positionAssignmentId, formVersionId, businessDate, payload]
      properties:
        orgUnitId: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        positionAssignmentId: { type: string, format: uuid }
        formVersionId: { type: string, format: uuid }
        businessDate: { type: string, format: date }
        payload: { type: object, additionalProperties: true }
        workPackageVersionId: { type: [string, 'null'], format: uuid }
        workPackageItemId: { type: [string, 'null'], format: uuid }
        workExpectationId: { type: [string, 'null'], format: uuid }
        recordKind: { type: [string, 'null'] }
        targetOrgUnitId: { type: [string, 'null'], format: uuid }
        occurredAt: { type: [string, 'null'], format: date-time }
        supersedesWorkRecordId: { type: [string, 'null'], format: uuid }
        saveAsDraft: { type: [boolean, 'null'] }
        completionStatement: { type: [string, 'null'] }
        exceptionStatement: { type: [string, 'null'] }
        nextAction: { type: [string, 'null'] }
    UpdateWorkRecordDraft:
      type: object
      required: [payload, expectedVersion]
      properties:
        payload: { type: object, additionalProperties: true }
        occurredAt: { type: [string, 'null'], format: date-time }
        completionStatement: { type: [string, 'null'] }
        exceptionStatement: { type: [string, 'null'] }
        nextAction: { type: [string, 'null'] }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    AddWorkRecordSupplement:
      type: object
      required: [submittedByAssignmentId, content]
      properties:
        submittedByAssignmentId: { type: string, format: uuid }
        content: { type: string, minLength: 1 }
    SubmitWorkRecordDraft:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    ReviewWorkRecord:
      type: object
      required: [outcome, expectedVersion]
      properties:
        outcome: { type: string, enum: [APPROVED, REJECTED] }
        reason: { type: [string, 'null'] }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
      allOf:
        - if:
            properties:
              outcome: { const: REJECTED }
          then:
            required: [reason]
            properties:
              reason: { type: string, minLength: 1 }
    AddAttachment:
      type: object
      required: [originalName, mediaType, sizeBytes]
      properties:
        originalName: { type: string }
        mediaType: { type: string }
        sizeBytes: { type: integer, format: int64, minimum: 0 }
        sha256: { type: [string, 'null'], pattern: '^[a-fA-F0-9]{64}$' }
    UploadedAttachment:
      type: object
      required: [id, workRecordId, objectKey, originalName, mediaType, sizeBytes, sha256, scanStatus, contentUrl]
      properties:
        id: { type: string, format: uuid }
        workRecordId: { type: string, format: uuid }
        objectKey: { type: string }
        originalName: { type: string }
        mediaType: { type: string, enum: [image/jpeg, image/png] }
        sizeBytes: { type: integer, format: int64, minimum: 1 }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        scanStatus:
          type: string
          enum: [CLEAN, BYPASSED_DEV]
          description: >-
            CLEAN is returned only after a configured scanner succeeds. BYPASSED_DEV is
            restricted to explicitly enabled development-header mode and is never accepted
            for download in formal mode.
        contentUrl: { type: string, pattern: '^/api/v1/work-data/attachments/.+/content$' }
    StoredAttachment:
      type: object
      required: [id, work_record_id, object_key, original_name, media_type, size_bytes, scan_status, created_at]
      properties:
        id: { type: string, format: uuid }
        work_record_id: { type: string, format: uuid }
        object_key: { type: string }
        original_name: { type: string }
        media_type: { type: string }
        size_bytes: { type: integer, format: int64, minimum: 0 }
        sha256: { type: [string, 'null'], pattern: '^[a-fA-F0-9]{64}$' }
        scan_status:
          type: string
          enum: [PENDING, CLEAN, BYPASSED_DEV]
          description: >-
            PENDING is retained for the deprecated metadata-only endpoint. Binary uploads
            are CLEAN only after a real scanner succeeds, or BYPASSED_DEV only in explicitly
            enabled development-header mode. Formal-mode downloads allow CLEAN only.
        created_at: { type: string, format: date-time }
...
    CreateEnterpriseTemplate:
      type: object
      required: [templateType, code, name, configuration]
      properties:
        templateType: { type: string, enum: [TASK, HOTEL_DASHBOARD] }
        code: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        targetPositionId: { type: [string, 'null'], format: uuid }
        ownerOrgUnitId: { type: [string, 'null'], format: uuid }
        configuration: { type: object, additionalProperties: true }
    CreateEnterpriseTemplateVersion:
      type: object
      required: [configuration]
      properties:
        configuration: { type: object, additionalProperties: true }
    UpdateEnterpriseTemplateVersion:
      type: object
      required: [configuration, expectedVersion]
      properties:
        configuration: { type: object, additionalProperties: true }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    PublishEnterpriseTemplateVersion:
      type: object
      properties:
        effectiveFrom: { type: [string, 'null'], format: date-time }
...